requirements.txt
Django==1.9.2
MySQL-python==1.2.5
PyJWT==1.4.0
argparse==1.2.1
django-braces==1.8.1
django-oauth-toolkit==0.10.0
django-rest-framework-social-oauth2==1.0.2
djangorestframework==3.3.2
oauthlib==1.0.3
python-openid==2.2.5
python-social-auth==0.2.14
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
wsgiref==0.1.2
